+++ Database/Entities/Prices.cs
@@ -7,7 +7,7 @@ namespace Database.Entities
         public int plid { get; set; }
         public string dps { get; set; }
         public string dpe { get; set; }
-        public float price { get; set; }
+        public decimal price { get; set; }
         public bool ispt { get; set; }
         public string days { get; set; }
     }
